pyproject.toml
[tool.poetry]
name = "pipeline-runner"
version = "0.1.14"
description = "Run a bitbucket pipeline locally"
authors = ["Mathieu Lemay <mathieu.lemay@broadsign.com>"]

[tool.poetry.dependencies]
python = "^3.6.1"
PyYAML = "^5.4.1"
docker = "^4.4.1"
click = ">=6.7,<8.0"
python-slugify = "^4.0.1"
boto3 = "^1.16.63"
GitPython = "^3.1.12"
python-dotenv = "^0.15.0"
xdg = "^5.0.1"
requests = "^2.25.1"
coloredlogs = "^15.0"
pyfzf = "^0.2.2"
beautifulsoup4 = "^4.9.3"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-mock = "^3.5.1"
tenacity = "^7.0.0"
coverage = "^5.5"

[tool.poetry.scripts]
pipeline-runner = 'pipeline_runner.cli:main'

[tool.black]
line-length = 120

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
